+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## [x]
+- Fix typo in `Formatter`
+
 ## [0.21.10]
 - Remove not needed `getLimitEnd` and rename `getLimitStart` to `getOffset`