+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [x]
 - Fix typo in `Formatter`
+- Include some formatters
 
 ## [0.21.10]
 - Remove not needed `getLimitEnd` and rename `getLimitStart` to `getOffset`