+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [x]
 - Fix typo in `Formatter`
 - Include some formatters
+- Add `StaticDataFetcher`
 
 ## [0.21.10]
 - Remove not needed `getLimitEnd` and rename `getLimitStart` to `getOffset`